TIP-603: extract attribute keys generation business
