+++ Resources/config/generators.yml
@@ -113,6 +113,8 @@ services:
             - '@pim_data_generator.product.product_raw_builder'
             - '@pim_catalog.repository.family'
             - '@pim_catalog.repository.group'
+            - '@pim_catalog.repository.attribute'
+            - '@pim_data_generator.attribute_key_provider'
         tags:
             - { name: pim_data_generator.generator }
